Step 1 of refactoring HomeController Account method that uses the Discord API (APP NOT TESTED WITH THESE CHANGES)
