+++ sprint_1/ProjectDiscordStat/DiscordStat/DAL/Abstract/IDiscordService.cs
@@ -1,0 +1,9 @@
+using DiscordStats.Models;
+
+namespace DiscordStats.DAL.Abstract
+{
+    public interface IDiscordService
+    {
+        public IEnumerable<Server>? GetCurrentUserGuilds(string bearerToken);
+    }
+}
