+++ sprint_1/ProjectDiscordStat/DiscordStat/Views/Home/Account.cshtml
@@ -5,9 +5,9 @@
 <ul>
 @foreach (var s in Model)
 {
-    <li>@s.owner</li>
-    <li>@s.name</li>
-    <li>@s.id</li>
+    <li>@s.Owner</li>
+    <li>@s.Name</li>
+    <li>@s.Id</li>
 
 }
 </ul>
